refactor(parsegamecode): deprecate: game_code in GameUS
The API no longer returns game_code for American games so the function now throws a
DeprecationWarning using the warning event, which will be printed in stderr. The function will
return an empty string whenever no game_code is present, or return the proper game_code if the API
decides to include it again. game_code has been marked as potentially undefined in the TypeScript
typings.

fixes #99