feat: add getQueriedGamesAmerica
You can use this to get games based on a query

BREAKING CHANGE: use getQueriedGamesAmerica to get games based on a given query. For example when
your users can perform a search (just like on the nintendo.com website) you can use this to severly
limit the results to go through.